chore: reorganize project structure
