+++ requirements.txt
@@ -12,18 +12,18 @@ mpmath==1.3.0
 networkx==3.2.1
 nltk==3.8.1
 numpy==1.26.2
-nvidia-cublas-cu12==12.1.3.1
-nvidia-cuda-cupti-cu12==12.1.105
-nvidia-cuda-nvrtc-cu12==12.1.105
-nvidia-cuda-runtime-cu12==12.1.105
-nvidia-cudnn-cu12==8.9.2.26
-nvidia-cufft-cu12==11.0.2.54
-nvidia-curand-cu12==10.3.2.106
-nvidia-cusolver-cu12==11.4.5.107
-nvidia-cusparse-cu12==12.1.0.106
-nvidia-nccl-cu12==2.18.1
-nvidia-nvjitlink-cu12==12.3.101
-nvidia-nvtx-cu12==12.1.105
+# nvidia-cublas-cu12==12.1.3.1
+# nvidia-cuda-cupti-cu12==12.1.105
+# nvidia-cuda-nvrtc-cu12==12.1.105
+# nvidia-cuda-runtime-cu12==12.1.105
+# nvidia-cudnn-cu12==8.9.2.26
+# nvidia-cufft-cu12==11.0.2.54
+# nvidia-curand-cu12==10.3.2.106
+# nvidia-cusolver-cu12==11.4.5.107
+# nvidia-cusparse-cu12==12.1.0.106
+# nvidia-nccl-cu12==2.18.1
+# nvidia-nvjitlink-cu12==12.3.101
+# nvidia-nvtx-cu12==12.1.105
 packaging==23.2
 Pillow==10.1.0
 PyYAML==6.0.1
@@ -41,10 +41,11 @@ torch==2.1.1
 torchvision==0.16.1
 tqdm==4.66.1
 transformers==4.35.2
-triton==2.1.0
+# triton==2.1.0
 typing_extensions==4.8.0
 urllib3==2.1.0
 python-dotenv==1.0.1
 vec2text==0.0.11
 fire==0.5.0
 loguru==0.7.2
+-e .
