+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(
+    name="tml-decoder",
+    version="0.1",
+    packages=["tml_decoder"],
+)
