+++ tml_decoder/models/abstract_model.py
